+++ src/App.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import { UserProvider } from './context/UserContext';
 import { SearchProvider } from './context/SearchContext';
 import HomePage from './pages/HomePage';
 import GlobalStyles from './GlobalStyles';
@@ -21,32 +22,34 @@ import ForumPage from './pages/ForumPage';
 
 function App() {
   return (
-    <SearchProvider>
-      <Router>
+    <UserProvider>
+      <SearchProvider>
+        <Router>
             <GlobalStyles />
-              <Routes>
-                  <Route path="" element={<HomePage />} />
-                  <Route path="/subcategories" element={<Subcategory/>}></Route>
-                  <Route path="/create" element={<CreatePost/>}></Route>
-                  <Route path="/user" element={<UserPage/>} />
-                  <Route path="/forum" element={<ForumPage/>} />
+            <Routes>
+                <Route path="" element={<HomePage />} />
+                <Route path="/subcategories" element={<Subcategory/>}></Route>
+                <Route path="/create" element={<CreatePost/>}></Route>
+                <Route path="/user" element={<UserPage/>} />
+                <Route path="/forum" element={<ForumPage/>} />
 
-                  <Route path="/:pageName" element={<FooterPage />} />
+                <Route path="/:pageName" element={<FooterPage />} />
 
-                  <Route path="/contact" element={<ContactPage/>} />
-                  <Route path="/help-faq-abuse-legal" element={<HelpFaqAbuseLegalPage/>} />
-                  <Route path="/avoid-scams" element={<AvoidScamsPage/>} />
-                  <Route path="/safety-tips" element={<PersonalSafetyTipsPage/>} />
-                  <Route path="/craigslist-about" element={<AboutCraigslistPage/>} />
-                  <Route path="/best-of-craigslist" element={<BestOfCraigslistPage/>} />
-                  <Route path="/careers" element={<CareersPage/>} />
-                  <Route path="/whats-new" element={<WhatsNewPage/>} />
-                  <Route path="/system-status" element={<SystemStatusPage/>} />
-                  <Route path="/charity" element={<SystemStatusPage/>} />
-                  <Route path="/philanthropies" element={<SystemStatusPage/>} />
-              </Routes>
-      </Router>
-    </SearchProvider>
+                <Route path="/contact" element={<ContactPage/>} />
+                <Route path="/help-faq-abuse-legal" element={<HelpFaqAbuseLegalPage/>} />
+                <Route path="/avoid-scams" element={<AvoidScamsPage/>} />
+                <Route path="/safety-tips" element={<PersonalSafetyTipsPage/>} />
+                <Route path="/craigslist-about" element={<AboutCraigslistPage/>} />
+                <Route path="/best-of-craigslist" element={<BestOfCraigslistPage/>} />
+                <Route path="/careers" element={<CareersPage/>} />
+                <Route path="/whats-new" element={<WhatsNewPage/>} />
+                <Route path="/system-status" element={<SystemStatusPage/>} />
+                <Route path="/charity" element={<SystemStatusPage/>} />
+                <Route path="/philanthropies" element={<SystemStatusPage/>} />
+            </Routes>
+        </Router>
+      </SearchProvider>
+    </UserProvider>
   );
 }
